third_party/WebKit/LayoutTests/editing/spelling/context_click_select_misspelling.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script src="spellcheck_test.js"></script>

<script>
test(() => assert_not_equals(window.internals, undefined),
     'This test requires internals to inspect INPUT elements.');

test(() => assert_not_equals(window.eventSender, undefined),
     'This test requires event sender to simulate keyboard and mouse actions.');

const kNumTests = 1;
const heldTest = async_test(
    () => assert_true(true),
    'Dummy async test for blocking test harness.',
    {finishCount: 0});

add_result_callback(testObj => {
  if (!testObj.properties.blockHeldTest)
    return;
  if (++heldTest.properties.finishCount === kNumTests)
    heldTest.done();
});

function findTextNode(node) {
  if (node.nodeName !== 'INPUT' && node.nodeName !== 'TEXTAREA')
    return node.firstChild;
  return internals.shadowRoot(node).getElementById('inner-editor').firstChild;
}

function assertContextClickSelection(container, offset, expected, title) {
  const document = container.ownerDocument;
  const textNode = findTextNode(container);
  const range = document.createRange();
  range.setStart(textNode, offset);

  const rect = range.getClientRects()[0];
  const x = document.offsetLeft + rect.left;
  const y = document.offsetTop + rect.top + rect.height / 2;
  eventSender.mouseMoveTo(x, y);
  eventSender.contextClick();

  // Esc key to hide the context menu.
  eventSender.keyDown('Escape', null);

  assert_equals(document.getSelection().toString(), expected, title);
}

function isMac(platform) {
  return platform.includes('Mac');
}

spellcheck_test(
    '<input type="text" value="wellcome home.">',
    document => document.querySelector('input').focus(),
    '<input type="text" value="#wellcome# home.">',
    {
      title: 'Mark misspelling in the initial text of INPUT.',
      callback: sample => test(() => {
        const container = sample.document.querySelector('input');
        assertContextClickSelection(
            container, 4, 'wellcome',
            'Context clicking "wellcome" selects the misspelled word');

        const shouldSelect = isMac(navigator.platform);
        if (shouldSelect) {
          assertContextClickSelection(
              container, 11, 'home',
              'Context clicking "home" selects the correctly spelled word');
          return;
        }
        assertContextClickSelection(
            container, 11, '',
            'Context clicking "home" does not select the correctly spelled word');
      },
      'Context clicking misspelled word in INPUT selects the word.',
      {blockHeldTest: true})
    });
</script>
